Make create idempotent. Fixes #4.
The create call was never checking to see if there was an existing
state for an instance. We now assume if kitchen is tracking an
instance we do not need to actually create a new server instance. This
has one feature flaw, which is if the instance has been deleted and the
statefile indicates it is created we will fail to communicate with the
instance rather than try to create another. This is a somewhat
intentional behavior as most of the kitchen drivers currently seem to
behave in this manner.